Started on rotary_table

+++ lib/ssd1306/ssd1306.hpp
@@ -8,7 +8,7 @@ class ssd1306 {
 private:
   twi_master *_twi;
 public:
-  ssd1306(uint8_t address, double freq, double clock);
+  ssd1306(uint8_t address, uint32_t freq, uint32_t clock);
   void write(uint8_t data);
 };
